moved services/api/api.js to services/api.js
