+++ frontend/src/pages/Login.js
@@ -3,7 +3,7 @@ import loginGreen from '../images/login-green.png';
 import loginBlue from '../images/login-blue.png';
 import styles from '../styles/Login.module.css';
 import { Link, withRouter } from 'react-router-dom';
-import api from '../services/api/api';
+import api from '../services/api';
 
 class Login extends Component {
     constructor(props) {
